Adjust to recent CGV Framework changes

+++ exercise2/implicit_group.cxx
@@ -6,7 +6,7 @@ template <typename T>
 void implicit_group<T>::set_update_handler(scene_update_handler* uh)
 {
 	implicit_base<T>::set_update_handler(uh);
-	for (size_t i = 0; i < get_nr_children(); ++i)
+	for (unsigned i = 0; i < get_nr_children(); ++i)
 		get_implicit_child(i)->set_update_handler(uh);
 }
 
@@ -43,7 +43,7 @@ const implicit_base<T>* implicit_group<T>::get_implicit_child(unsigned i) const
 template <typename T>
 void implicit_group<T>::on_set(void* member_ptr)
 {
-	for (size_t i = 0; i < group::get_nr_children(); ++i) {
+	for (unsigned i = 0; i < group::get_nr_children(); ++i) {
 		bool& toggle = provider::ref_tree_node_visible_flag(*get_implicit_child(i));
 		if (member_ptr == &child_visible_in_gui[i]) {
 			toggle = (bool&)child_visible_in_gui[i];
@@ -64,7 +64,7 @@ void implicit_group<T>::on_set(void* member_ptr)
 template <typename T>
 unsigned int implicit_group<T>::append_child(base_ptr child)
 {
-	size_t i = group::append_child(child);
+	unsigned i = group::append_child(child);
 	child_visible_in_gui.push_back(1);
 	return i;
 }
@@ -73,13 +73,13 @@ unsigned int implicit_group<T>::append_child(base_ptr child)
 template <typename T>
 typename implicit_group<T>::clr_type implicit_group<T>::compose_color(const pnt_type& p) const
 {
-	size_t n = group::get_nr_children();
+	unsigned n = group::get_nr_children();
 	switch (n) {
 	case 0: return implicit_base<T>::color;
 	case 1: return get_implicit_child(0)->evaluate_color(p);
 	default: {
 		clr_type result(0, 0, 0, 0);
-		for (size_t i = 0; i < n; ++i)
+		for (unsigned i=0; i<n; ++i)
 			result += 1.0f / float(n) * get_implicit_child(i)->evaluate_color(p);
 		return result;
 	}
@@ -108,7 +108,7 @@ template <typename T>
 void implicit_group<T>::create_gui()
 {
 	provider::align("\a");
-	for (size_t i = 0; i < get_nr_children(); ++i) {
+	for (unsigned i = 0; i < get_nr_children(); ++i) {
 		implicit_base<T>* child_ptr = get_implicit_child(i);
 		provider::ref_tree_node_visible_flag(*child_ptr) = (bool&)child_visible_in_gui[i];
 
@@ -141,7 +141,7 @@ template <typename T>
 bool implicit_group<T>::init(context& ctx)
 {
 	bool success = true;
-	for (size_t i = 0; i < group::get_nr_children(); ++i)
+	for (unsigned i = 0; i < group::get_nr_children(); ++i)
 		success = get_implicit_child(i)->init(ctx) && success;
 	return success;
 }
